+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - BaseDivider: Add margin prop.
 
-- BaseHeading: Add fontSize prop.
+- BaseHeading: Add fontSize and lineHeight prop.
 
 - FieldCheckbox: Add autofocus, error, info, success, validation and warning prop.